requirements.txt
attrs==22.1.0
beautifultable==1.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.0
filelock==3.8.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==2.4.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pysondb-v2==2.1.0
pytest==7.2.0
PyYAML==3.11
requests==2.28.1
termcolor==2.0.1
tomli==2.0.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
